console: MACROS to print colored text, errors, warnings elf: fix compiler warning

+++ libxenon/drivers/elf/elf.c
@@ -18,6 +18,7 @@ see file COPYING or http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt
 #include <libfdt/libfdt.h>
 #include <nocfe/addrspace.h>
 
+#include "elf.h"
 #include "elf_abi.h"
 
 #define INITRD_RELOC_START ((void*)0x85FE0000)
@@ -293,14 +294,14 @@ void elf_runWithDeviceTree (void *elf_addr, int elf_size, void *dt_addr, int dt_
                 kernel_relocate_initrd(initrd_start,initrd_size);
                 
                 u64 start, end;
-		start = (u32)PHYSADDR((uint64_t)initrd_start);
+		start = (u32)PHYSADDR((u32)initrd_start);
 		res = fdt_setprop(ELF_DEVTREE_START, node, "linux,initrd-start", &start, sizeof(start));
 		if (res < 0){
 			printf("couldn't set chosen.linux,initrd-start property\n");
                         return;
                 }
 
-		end = (u32)PHYSADDR((uint64_t)initrd_start + (uint64_t)initrd_size);
+		end = (u32)PHYSADDR(((u32)initrd_start + (u32)initrd_size));
 		res = fdt_setprop(ELF_DEVTREE_START, node, "linux,initrd-end", &end, sizeof(end));
 		if (res < 0) {
 			printf("couldn't set chosen.linux,initrd-end property\n");
@@ -363,7 +364,7 @@ void kernel_relocate_initrd(void *start, size_t size)
         initrd_size = size;
         
         printf("Initrd at %p/0x%lx: %ld bytes (%ldKiB)\n", initrd_start, \
-        (u32)PHYSADDR((uint64_t)initrd_start), initrd_size, initrd_size/1024);
+        (u32)PHYSADDR((u32)initrd_start), initrd_size, initrd_size/1024);
 }
 
 void kernel_reset_initrd(void)
